# openapi3_0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BlogLite - OpenAPI 3.0",
    "description": "MarkDown blogging site.",
    "contact": {
      "email": "21f1000263@student.onlinedegree.iitm.ac.in"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000/api/"
    }
  ],
...
  "paths": {
...
    "/entry": {
      "get": {
        "security": [
          {
            "api_key": [],
            "api_secret": []
          }
        ],
        "tags": [
          "Entry",
          "Retrive Entries"
        ],
        "summary": "Get an entry by post ID.",
        "operationId": "retrive_a_post",
        "parameters": [
          {
            "in": "query",
            "description": "Required for retriving exact entry.",
            "name": "pid",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An entry with corresponding informations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "integer",
                      "format": "int64",
                      "example": 4
                    },
                    "content": {
                      "type": "string"
                    },
                    "date": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "id": {
                      "type": "integer"
                    },
                    "liked": {
                      "type": "integer"
                    },
                    "likes": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "profile_pic": {
                      "type": "string"
                    },
                    "shared": {
                      "type": "integer"
                    },
                    "shares": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized Access Denied."
          },
          "404": {
            "description": "No entry with that id found."
          }
        }
      },
      "post": {
        "security": [
          {
            "api_key": [],
            "api_secret": []
          }
        ],
        "tags": [
          "Entry"
        ],
        "summary": "Create a new entry.",
        "description": "This endpoint can be used to create a new entry.",
        "operationId": "create_entry",
        "parameters": [
          {
            "in": "query",
            "name": "content",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image file name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "integer",
                      "format": "int64",
                      "example": 4
                    },
                    "content": {
                      "type": "string"
                    },
                    "date": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "id": {
                      "type": "integer"
                    },
                    "liked": {
                      "type": "integer"
                    },
                    "likes": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "profile_pic": {
                      "type": "string"
                    },
                    "shared": {
                      "type": "integer"
                    },
                    "shares": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized Access Denied."
          }
        }
      },
      "delete": {
        "security": [
          {
            "api_key": [],
            "api_secret": []
          }
        ],
        "tags": [
          "Entry"
        ],
        "summary": "Delete an entry by its ID.",
        "operationId": "delete_entry",
        "parameters": [
          {
            "in": "query",
            "name": "pid",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entry deleted sucessfully."
          },
          "401": {
            "description": "Unauthorized Access Denied."
          },
          "404": {
            "description": "No entry with that id found."
          }
        }
      },
      "put": {
        "security": [
          {
            "api_key": [],
            "api_secret": []
          }
        ],
        "tags": [
          "Entry"
        ],
        "summary": "Update an entry.",
        "operationId": "update_entry",
        "parameters": [
          {
            "in": "query",
            "name": "pid",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "content",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entry updated sucessfully."
          },
          "401": {
            "description": "Unauthorized Access Denied."
          },
          "404": {
            "description": "No entry with that id found."
          }
        }
      }
    },
...
    "/entry/username": {
      "get": {
        "security": [
          {
            "api_key": [],
            "api_secret": []
          }
        ],
        "tags": [
          "Retrive Entries"
        ],
        "summary": "Get an entry by post ID.",
        "description": "Only one parameter is expected at a time.",
        "operationId": "users_all_post",
        "parameters": [
          {
            "in": "query",
            "required": true,
            "name": "username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An entry with corresponding informations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "integer",
                      "format": "int64",
                      "example": 4
                    },
                    "content": {
                      "type": "string"
                    },
                    "date": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "id": {
                      "type": "integer"
                    },
                    "liked": {
                      "type": "integer"
                    },
                    "likes": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "profile_pic": {
                      "type": "string"
                    },
                    "shared": {
                      "type": "integer"
                    },
                    "shares": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No username with that username found."
          }
        }
      }
    },
...
    "/entry/trending": {
      "get": {
        "security": [
          {
            "api_key": [],
            "api_secret": []
          }
        ],
        "tags": [
          "Retrive Entries"
        ],
        "summary": "Get trending entries.",
        "operationId": "trending",
        "responses": {
          "200": {
            "description": "An entry with corresponding informations.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "comments": {
                      "type": "integer",
                      "format": "int64",
                      "example": 4
                    },
                    "content": {
                      "type": "string"
                    },
                    "date": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "id": {
                      "type": "integer"
                    },
                    "liked": {
                      "type": "integer"
                    },
                    "likes": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "profile_pic": {
                      "type": "string"
                    },
                    "shared": {
                      "type": "integer"
                    },
                    "shares": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/entry/archive": {
      "patch": {
        "security": [
          {
            "api_key": [],
            "api_secret": []
          }
        ],
        "tags": [
          "Intraction"
        ],
        "summary": "Archive a post to unpublish it.",
        "operationId": "archive",
        "parameters": [
          {
            "in": "query",
            "name": "pid",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entry archived sucessfully."
          },
          "401": {
            "description": "Unauthorized Access Denied."
          },
          "404": {
            "description": "No entry with that id found."
          }
        }
      }
    },
...
    "/user/followings": {
      "patch": {
        "security": [
          {
            "api_key": [],
            "api_secret": []
          }
        ],
        "tags": [
          "Stats"
        ],
        "summary": "Retrive List of Followings",
        "operationId": "followings",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user being followed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile_pic": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/user/followers": {
      "patch": {
        "security": [
          {
            "api_key": [],
            "api_secret": []
          }
        ],
        "tags": [
          "Stats"
        ],
        "summary": "Retrive List of Followers.",
        "operationId": "followers",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of user being following current user..",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profile_pic": {
                      "type": "string"
                    },
                    "username": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/user/is_available": {
      "patch": {
        "security": [
          {
            "api_key": [],
            "api_secret": []
          }
        ],
        "tags": [
          "Stats"
        ],
        "summary": "Check that username is being used or not.",
        "operationId": "is_available",
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Username is being used."
          },
          "417": {
            "description": "Username is available"
          }
        }
      }
    },
    "/user/is_following": {
      "patch": {
        "security": [
          {
            "api_key": [],
            "api_secret": []
          }
        ],
        "tags": [
          "Stats"
        ],
        "summary": "Check if user is following current_user or not.",
        "operationId": "is_following",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "//Boolean// User is following?",
            "content": {
              "plain/text": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "417": {
            "description": "Username is not available"
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      },
      "api_secret": {
        "type": "apiKey",
        "name": "api_secret",
        "in": "header"
      }
    }
  }
}